[feature] add design tokens and layout
