+++ glancy-site/src/index.css
@@ -1,7 +1,10 @@
 @import './theme/colors.css';
+@import './theme/spacing.css';
+@import './theme/typography.css';
+@import './theme/layout.css';
 
 :root {
-  font-family: system-ui, Avenir, Helvetica, Arial, sans-serif;
+  font-family: var(--font-family-base);
   line-height: 1.5;
   font-weight: 400;
