+++ glancy-site/src/theme/colors.css
@@ -16,4 +16,12 @@
 
   --accent-color: #facc15;
   --brand-color: orange;
+
+  /* surface tones */
+  --surface-light: #f7f7f7;
+  --surface-dark: #1b1b1b;
+
+  /* text scales */
+  --text-primary-light: #1a1a1a;
+  --text-primary-dark: #f5f5f5;
 }
